% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysis.R
\name{doAddtionalMetaAnalysis}
\alias{doAddtionalMetaAnalysis}
\title{Additional meta-analysis}
\usage{
doAddtionalMetaAnalysis(
  shinyFolder,
  maExportFolder,
  maxCores,
  interactions = F,
  positiveControlOutcome = FALSE,
  databaseId = "Meta-analysis"
)
}
\arguments{
\item{shinyFolder}{A folder name containing shiny result. make sure to use forward slashes (/). D}

\item{maExportFolder}{A local folder where the meta-anlysis results will be written.}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}

\item{databaseId}{desginated databaseId (default = 'Meta-analysis')}
}
\description{
Additional meta-analysis
}
\details{
This function generates tables and figures for the report on the study results.
}
