% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysis.R
\name{doMetaAnalysis}
\alias{doMetaAnalysis}
\title{Create figures and tables for report}
\usage{
doMetaAnalysis(
  shinyFolder,
  maExportFolder,
  maxCores,
  interactions,
  positiveControlOutcome = FALSE,
  diagnostics = NULL,
  databaseId = "Meta-analysis"
)
}
\arguments{
\item{shinyFolder}{A folder name containing shiny result. make sure to use forward slashes (/). D}

\item{maExportFolder}{A local folder where the meta-analysis results will be written.}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}

\item{databaseId}{}
}
\description{
Create figures and tables for report
}
\details{
This function generates tables and figures for the report on the study results.
}
